Move particle_functor to its own submodul and general clean-up
