+++ lib/python/picongpu/picmi/diagnostics/particle_dump.py
@@ -10,11 +10,10 @@ from pathlib import Path
 
 from pydantic import BaseModel
 
-from picongpu.picmi.species import Species
+from picongpu.picmi.diagnostics.backend_config import BackendConfig, OpenPMDConfig
+from picongpu.picmi.diagnostics.timestepspec import TimeStepSpec
 from picongpu.picmi.particle_functor.particle_filter import FilteredSpecies
-
-from .backend_config import BackendConfig, OpenPMDConfig
-from .timestepspec import TimeStepSpec
+from picongpu.picmi.species import Species
 
 
 class ParticleDump(BaseModel):
