+++ lib/python/picongpu/pypicongpu/output/__init__.py
@@ -5,14 +5,14 @@ Authors: Julian Lenz
 License: GPLv3+
 """
 
-from .phase_space import PhaseSpace
+from .checkpoint import Checkpoint
 from .energy_histogram import EnergyHistogram
 from .macro_particle_count import MacroParticleCount
-from .timestepspec import TimeStepSpec
-from .checkpoint import Checkpoint
 from .openpmd_plugin import OpenPMDPlugin
+from .phase_space import PhaseSpace
 from .plugin import Plugin
 from .radiation import RadiationConfiguration, RadiationPlugin, RadiationObserverConfiguration
+from .timestepspec import TimeStepSpec
 
 __all__ = [
     "OpenPMDPlugin",
