+++ lib/python/picongpu/pypicongpu/output/macro_particle_count.py
@@ -5,11 +5,11 @@ Authors: Masoud Afshari, Julian Lenz
 License: GPLv3+
 """
 
-from .timestepspec import TimeStepSpec
 from pydantic import BaseModel, PrivateAttr
-from ..species import Species
 
-from .plugin import Plugin
+from picongpu.pypicongpu.output.plugin import Plugin
+from picongpu.pypicongpu.output.timestepspec import TimeStepSpec
+from picongpu.pypicongpu.species import Species
 
 
 class MacroParticleCount(Plugin, BaseModel):
